+++ changelog.md
@@ -2,3 +2,4 @@
 - add option for dictionary to take hunger as well as XP
 - add config GUI
 - add config options for hardcore sacrifice and activation
+- made pattern finder ignore active runes